Add files via upload
Given an array of N integers. Find the first element that occurs at least K number of times.
Task is to complete the function firstElementKTime() which takes the array A[], its size N, and an integer K as inputs and returns the required answer. 
If the answer is not present in the array, return -1.